cred_ssh_key {git2r}	R Documentation

Create a new passphrase-protected ssh key credential object

Description

Create a new passphrase-protected ssh key credential object

Usage

cred_ssh_key(publickey = ssh_path("id_rsa.pub"),
  privatekey = ssh_path("id_rsa"), passphrase = character(0))

Arguments

publickey
The path to the public key of the credential. Default is ssh_path("id_rsa.pub")

privatekey
The path to the private key of the credential. Default is ssh_path("id_rsa")

passphrase
The passphrase of the credential. Default is character(0). If getPass is installed and private key is passphrase protected getPass::getPass() will be called to allow for interactive and obfuscated interactive input of the passphrase.

Value

A list of class cred_ssh_key with entries:

publickey

The path to the public key of the credential

privatekey

The path to the private key of the credential

passphrase

The passphrase of the credential

See Also

Other git credential functions: cred_env, cred_token, cred_user_pass

Examples

## Not run: 
## Create a ssh key credential object. It can optionally be
## passphrase-protected
cred <- cred_ssh_key(ssh_path("id_rsa.pub"), ssh_path("id_rsa"))
repo <- repository("git2r")
push(repo, credentials = cred)

## End(Not run)

[Package git2r version 0.25.2 Index]