cred_user_pass {git2r}	R Documentation

Create a new plain-text username and password credential object

Description

Create a new plain-text username and password credential object

Usage

cred_user_pass(username = NULL, password = NULL)

Arguments

username
The username of the credential

password
The password of the credential. If getPass is installed and the only input is username, getPass::getPass() will be called to allow for interactive and obfuscated interactive input of the password.

Value

A list of class cred_user_pass with entries:

username

The username of the credential

password

The password of the credential

See Also

Other git credential functions: cred_env, cred_ssh_key, cred_token

Examples

## Not run: 
## Create a plain-text username and password credential object
cred_user_pass("Random Developer", "SecretPassword")

## End(Not run)

[Package git2r version 0.25.2 Index]