note_default_ref {git2r}	R Documentation

Default notes reference

Description

Get the default notes reference for a repository

Usage

note_default_ref(repo = ".")

Arguments

repo
a path to a repository or a git_repository object. Default is '.'

Value

Character vector of length one with name of default notes reference

Examples

## Not run: 
## Create and initialize a repository in a temporary directory
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)
config(repo, user.name="Alice", user.email="alice@example.org")

## View default notes reference
note_default_ref(repo)

## End(Not run)

[Package git2r version 0.25.2 Index]