remote_ls {git2r}	R Documentation

List references in a remote repository

Description

Displays references available in a remote repository along with the associated commit IDs. Akin to the 'git ls-remote' command.

Usage

remote_ls(name = NULL, repo = NULL, credentials = NULL)

Arguments

name
Character vector with the "remote" repository URL to query or the name of the remote if a repo argument is given.

repo
an optional repository object used if remotes are specified by name.

credentials
The credentials for remote repository access. Default is NULL. To use and query an ssh-agent for the ssh key credentials, let this parameter be NULL (the default).

Value

Character vector for each reference with the associated commit IDs.

Examples

## Not run: 
remote_ls("https://github.com/ropensci/git2r")

## End(Not run)

[Package git2r version 0.25.2 Index]