remote_url {git2r}	R Documentation

Get the remote url for remotes in a repo

Description

Get the remote url for remotes in a repo

Usage

remote_url(repo = ".", remote = NULL)

Arguments

repo
a path to a repository or a git_repository object. Default is '.'

remote
Character vector with the remotes to get the url from. Default is the remotes of the repository.

Value

Character vector with remote_url for each of the remote

Examples

## Not run: 
## Initialize a temporary repository
path <- tempfile(pattern="git2r-")
dir.create(path)
repo <- init(path)

## Create a user and commit a file
config(repo, user.name="Alice", user.email="alice@example.org")
writeLines("Hello world!", file.path(path, "example.txt"))
add(repo, "example.txt")
commit(repo, "First commit message")

## Add a remote
remote_add(repo, "playground", "https://example.org/git2r/playground")
remotes(repo)
remote_url(repo, "playground")

## Rename a remote
remote_rename(repo, "playground", "foobar")
remotes(repo)
remote_url(repo, "foobar")

## Set remote url
remote_set_url(repo, "foobar", "https://example.org/git2r/foobar")
remotes(repo)
remote_url(repo, "foobar")

## Remove a remote
remote_remove(repo, "foobar")
remotes(repo)

## End(Not run)

[Package git2r version 0.25.2 Index]