revparse_single {git2r}	R Documentation

Revparse

Description

Find object specified by revision.

Usage

revparse_single(repo = ".", revision = NULL)

Arguments

repo
a path to a repository or a git_repository object. Default is '.'

revision
The revision string, see http://git-scm.com/docs/git-rev-parse.html#_specifying_revisions

Value

a git_commit or git_tag or git_tree object

Examples

## Not run: 
## Create a directory in tempdir
path <- tempfile(pattern="git2r-")
dir.create(path)

## Initialize a repository
repo <- init(path)
config(repo, user.name="Alice", user.email="alice@example.org")

## Create a file, add and commit
writeLines("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
           con = file.path(path, "test.txt"))
add(repo, "test.txt")
commit(repo, "First commit message")

# Change file and commit
writeLines(c("Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do",
             "eiusmod tempor incididunt ut labore et dolore magna aliqua."),
             con = file.path(path, "test.txt"))
add(repo, "test.txt")
commit(repo, "Second commit message")

revparse_single(repo, "HEAD^")
revparse_single(repo, "HEAD:test.txt")

## End(Not run)

[Package git2r version 0.25.2 Index]