ssl_cert_locations {git2r}	R Documentation

Set the SSL certificate-authority locations

Description

Set the SSL certificate-authority locations

Usage

ssl_cert_locations(filename = NULL, path = NULL)

Arguments

filename
Location of a file containing several certificates concatenated together. Default NULL.

path
Location of a directory holding several certificates, one per file. Default NULL.

Value

invisible(NULL)

Note

Either parameter may be 'NULL', but not both.

[Package git2r version 0.25.2 Index]